+++ crates/chainlink-datastreams/src/error.rs
@@ -10,4 +10,7 @@ pub enum Error {
     /// Overflow.
     #[error("overflow: {0}")]
     Overflow(&'static str),
+    /// Unknown market status.
+    #[error("unknown market status")]
+    UnknownMarketStatus,
 }