+++ programs/store/src/instructions/oracle/custom.rs
@@ -154,7 +154,9 @@ impl UpdatePriceFeedWithChainlink<'_> {
             msg!("Invalid report: {}", err);
             let err = match err {
                 ChainlinkError::NegativePrice(_) => CoreError::NegativePriceIsNotSupported,
-                ChainlinkError::InvalidRange(_) => CoreError::InvalidPriceReport,
+                ChainlinkError::InvalidRange(_) | ChainlinkError::UnknownMarketStatus => {
+                    CoreError::InvalidPriceReport
+                }
                 ChainlinkError::Overflow(_) => CoreError::PriceOverflow,
             };
             error!(err)