+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - program(store): Added instructions to configure token metadata.
 - program: Added `security.txt` to all public programs.
 - sdk(sdk): Added support for the instructions to configure token metadata.
+- sdk(chainlink-datastreams): Added `market_status` field and corresponding method to the `Report` struct.
+- sdk(chainlink-datastreams): Added support for the report schema v2 and the RWA report schema (v4) for Chainlink Data Streams.
 
 ### Changed
 - sdk: Enabled oracle price updates to be posted in parallel.